fix: move MockConfigLoader to domain/config/configtest
Move MockConfigLoader from internal/agent/agenttest (Application layer)
to internal/domain/config/configtest (Domain layer) where it belongs
since it mocks domain/config.ConfigLoader - a Domain interface.

Also fix 4 agent-layer test files that imported infra_config for
NewNoOpConfigWatcher/NewFileConfigWatcher factory functions. These now
use domain_config.NewNoOpConfigWatcher which already existed in
domain/config.

Resolves #264 - eliminates 3 Clean Architecture layer violations:
- agent -> infrastructure/config
- agent/agentinternal -> infrastructure/config
- infrastructure/config -> agent/agenttest

Zero production code changed. All ADRs (021, 022, 030) satisfied.
