+++ internal/domain/config/configtest/doc.go
@@ -1,0 +1,9 @@
+// Copyright (c) 2026 gosharplite@gmail.com
+// SPDX-License-Identifier: MIT
+
+// Package configtest provides test doubles for interfaces defined in
+// the parent internal/domain/config package.
+//
+// Types in this package are intended only for use from _test.go files.
+// Production code must never import this package.
+package configtest
