+++ internal/domain/config/configtest/mock_config_loader.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 gosharplite@gmail.com
 // SPDX-License-Identifier: MIT
 
-package agenttest
+package configtest
 
 import (
 	"github.com/gosharplite/tell-me-go/internal/domain/config"
